Restructure `shared` and `helper` folder
The readme explains what to put where.

This also extracts the OsmUrl shared function.
